Create Banned Users
For someone to end up on the banned users section they either have to be a notorious spammer or are promoting a potentially dangerous product/website. 

Due to the nature of this list, there is a name and URL to the channel and Google Plus profile. This way you can verify you are blocking the correct person.